Separate active address migration state
