+++ app/ts/background/settings.ts
@@ -13,6 +13,7 @@ import { DEFAULT_ACTIVE_ADDRESSES, DEFAULT_BLOCK_MANIPULATION, DEFAULT_RPCS } fr
 import { silenceChromeUnCaughtPromise } from '../utils/requests.js'
 import { mergeStoredWebsiteMetadata, sanitizeWebsiteAccess } from '../utils/websiteIcons.js'
 import type { SigningAddressPreference, SigningAddressPreferences } from '../types/signerTypes.js'
+import { createIndependentActiveSimulationAddressStorageUpdate } from './activeSimulationAddressStorage.js'
 
 export const defaultActiveAddresses = DEFAULT_ACTIVE_ADDRESSES
 
@@ -178,7 +179,7 @@ export async function changeSimulationMode(changes: { simulationMode: boolean, r
 	return await browserStorageLocalSet({
 		simulationMode: changes.simulationMode,
 		...changes.rpcNetwork ? { activeRpcNetwork: changes.rpcNetwork }: {},
-		...'activeSimulationAddress' in changes ? { activeSimulationAddress: changes.activeSimulationAddress, hasIndependentActiveSimulationAddress: true }: {},
+		...'activeSimulationAddress' in changes ? createIndependentActiveSimulationAddressStorageUpdate(changes.activeSimulationAddress) : {},
 		...'activeSigningAddress' in changes ? { activeSigningAddress: changes.activeSigningAddress }: {},
 		...'activeSigningSafeAddress' in changes ? { activeSigningSafeAddress: changes.activeSigningSafeAddress }: {},
 	})
